Move models for sub-resources into their own package
