+++ src/main/java/de/htwsaar/vs/chat/model/sub/Password.java
@@ -1,5 +1,6 @@
-package de.htwsaar.vs.chat.model;
+package de.htwsaar.vs.chat.model.sub;
 
+import de.htwsaar.vs.chat.model.User;
 import lombok.Data;
 
 import javax.validation.constraints.NotBlank;
