+++ src/main/java/de/htwsaar/vs/chat/model/sub/Role.java
@@ -1,12 +1,13 @@
-package de.htwsaar.vs.chat.model;
+package de.htwsaar.vs.chat.model.sub;
 
+import de.htwsaar.vs.chat.model.User;
 import lombok.Data;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Pattern;
 
 /**
- * Role Object Model.
+ * Role object model.
  *
  * @author Mahan Karimi
  * @see User
