+++ src/main/java/de/htwsaar/vs/chat/model/sub/Status.java
@@ -1,9 +1,10 @@
-package de.htwsaar.vs.chat.model;
+package de.htwsaar.vs.chat.model.sub;
 
+import de.htwsaar.vs.chat.model.User;
 import lombok.Data;
 
 /**
- * Status Object Model.
+ * Status object model.
  *
  * @author Mahan Karimi
  * @see User
